Updated openstack/openstack Project: openstack-infra/project-config  61d1c6943bd59e8fdcaf45ec4b86925fa0294ea8
Don't run check-requirements on old branches fro networking-bgppvn

We the check-requirements job was added to the template for
networking-bgppvn it enabled requirements checking on all branches.
Neither the networking-bgpvpn or the requirements repo is setup for this
on older (pre-Ocata) branches.  This is manifesting as updates to
requirements.txt to correct gate issues[1] being blocked by the
check-requirements job[2].

So disable that check job on old branches

[1] http://logs.openstack.org/periodic-stable/periodic-networking-bgpvpn-python27-newton/d91cbf9/console.html#_2017-03-01_06_12_03_108318
[2] http://logs.openstack.org/68/439668/2/check/gate-networking-bgpvpn-requirements/89d17d3/console.html#_2017-03-01_17_21_02_404982

Change-Id: I9b8364df78f19ca1f28b42ee24e715ef6abee33f
Needed-By: Ic109c11cbc72da4f206606995be156fc0f664266