Updated openstack/openstack Project: openstack/cinder  101dcc6ccafeaa317ec5131570e8e65c9505cfc8
Dell EMC SC: Added retry to _init_volume

To initialize a SC volume we quickly map and unmap the volume to
a scserver object. _init_volume currently tries to activate a
volume one time. If this fails we should retry until it works or
we run out of scserver objects to try.

Closes-Bug: #1672768
Change-Id: I383980a8241ec871313e6e7d7157f7af264fc5b6