Updated openstack/openstack Project: openstack-infra/nodepool  30c4b46b48aa89453531b3feda454473c98d845b
Default config-drive to true

As we depend more and more on glean to help bootstrap a node, it is
possible for new clouds added to nodepool.yaml to be missing the
setting. Which results is broken nodes and multiple configuration
updates.

As a result, we now default config-drive to true to make it easier to
bring nodes online.

Change-Id: I4e214ba7bc43a59ddffb4bfb50576ab3b96acf69
Signed-off-by: Paul Belanger <pabelanger@redhat.com>