Updated openstack/openstack Project: openstack/nova  e611116e5822abf0e9f933cf6bd1b4c63007b170
Make versioned notifications topics configurable

Some services (such as telemetry) actually consume the notifications. So
if one deploys a service that listens on the same queue as telemetry,
there will be race-conditions with these services and one will not get
the notifications that are expected at points.

To address this, one sets a different topic and consumes from there.
This is not possible with versioned notifications at the moment. And, as
services move to using that, the same need will arise.

So, this adds a configuration option to nova for enabling the
configuration of topics for this notifier.

Change-Id: I817ce4bae0dd37e0d06bd44f21ba81b3cb800548