Updated openstack/openstack Project: openstack-infra/tripleo-ci  f16697a32565bcfcaa8fed03ac70dcfb648e0406
oooq multinode: fix for stable branches

Current quickstart setup is not working with stable branches for the
multinode jobs.
Since we're still using the bootstrap script from tripleo.sh to
bootstrap nodes in multinode, we need to create and propagate deploy.env
to the subnodes.
We use existing functions, with some changes.

Partial-Bug: #1673051
Change-Id: Id05dcc64fcf7f67792400447a73ba179a9d4db4f