Updated openstack/openstack Project: openstack/cinder  f991cdc03d3a8c4ab7c5297e5a78523e37a6a1cc
Update devref for OVO version bumping

Our devref is not explicit enough on when a versioned object change
requires a version bump and this is creating confusion in developers and
reviewers.

This patch adds explicitly when bumps are required, when not, and how to
automatically check if it's required or not.

Change-Id: I4dd21c30b3a844ff47b3b3370ecd51d3e50655d5