Updated openstack/openstack Project: openstack-infra/project-config  3b8ffd42a63a3ce5e970c4d43745e54d6fdf6f74
Add neutron-dsvm-api job into devstack-gate check queue

The job was broken twice in the course of last two weeks (once by
local.conf loading strategy changed in d-g; now by incorrectly calling
testr slowest in the job).

And it's not surprising, since the job uses neutron gate_hook that is
very complex and relies on DEVSTACK_LOCALCONF, dsconf, and other gate
configuration features that may be not utilized in current list of
gating jobs.

The job is very stable, so there should have been no concerns about
destabilizing the d-g gate. But there are still concerns about tangling
d-g gate with external consuming projects too much. We will start with a
non-voting job, and see if it is enough to make neutron gate a safer
place.

Change-Id: I54d8023e11afec858cdceeb83cd2a57febc54971