Updated openstack/openstack Project: openstack/cinder  8c118bf502475c76c91aa162999e21d3cb0e9df3
Revert "Remove v1 API reference"

The v1 API is still in the tree, even if it is off by
default and deprecated.  We need to keep documentation on
how it works at least until we remove the code.  This was
published to [1] which was the primary way to find this
information, and it is now missing.

[1] https://developer.openstack.org/api-ref/block-storage/v1/

This marks v1 as deprecated in the toctree.

This reverts commit 8139f7f108786112ec6eef47228ecbc7f54b6f47.

Change-Id: Ic9d0fcce5ca0f8455c212292664b8f9694edfce3