Updated openstack/openstack Project: openstack-infra/tripleo-ci  21ec3361d363e38e758fa3da10ca7e4a9e8e99ce
Make sure we upgrade to the right repos during an upgrade

We currently call tripleo.sh --repo-setup during an upgrade without
setting STABLE_RELEASE. This will cause an upgrade to the latest master
which will break in a bunch of situations. E.g. Latest puppet-tripleo
without corresponding new THT package on the undercloud.

Change-Id: I712abe71f97c22ee3d55d9db2f641096f8a7350c
Closes-Bug: #1673030