Updated openstack/openstack Project: openstack-infra/devstack-gate  bd082d261a23ab2ddc6f3b1b0f470552beb2013e
Keep working to clean up PROJECTS list

This uses info provided to devstack-gate like whether or not Trove or
Sahara are enabled or which libs should be installed from source and not
pypi to populate a more minimal list of PROJECTS by default.

Note this doesn't switch all lib projects over to using the libs from
git signalling initially. Will test with a set of hopefully less common
libs to reduce impact.

Change-Id: Id2caf9561b361c1d4950856423282535d067e015