Updated openstack/openstack Project: openstack/nova  2380659e358770a3f36253b93a112b9779a23958
Add a functional regression/recreate test for bug 1671648

This adds a test which recreates the regression bug introduced
in Ocata where build retries are not populated when creating
instances in conductor for cells v2.

The change that fixes the bug will go on top of this and modify
the test to show the bug is fixed.

Change-Id: Ie9e955d79b4e1441092183135b3f70b003c94db5
Related-Bug: #1671648