Updated openstack/openstack Project: openstack/oslo.i18n  f294b2d5054f421ccf13e110b2cc67128d4bebd8
add notes about skipping log translation setup

Now that we have decided to skip log translations, let's stop telling
people to set them up. This change adds a note saying the work is not
needed, but leaves the instructions so that someone trying to understand
existing code can review them. The actual translation functions will be
marked for deprecation in the future, when enough projects have dropped
using them to warrant that change.

Change-Id: Ia34ad79ca9c6d1ea862b6954669c992a898a03ff
Signed-off-by: Doug Hellmann <doug@doughellmann.com>