Updated openstack/openstack Project: openstack-infra/irc-meetings  b701bf79e44be27dbdd4bcf01604bdbb59b05789
Update Common Classification Framework meeting

The time that was set previously was incompatible with most of the
people that have lately been engaged or interested in this effort.

This modifies the time to a more popular one amongst interested people
and also updates the name of the project (keeping IDs and unique refs).

Related message on the openstack-dev mailing list:
http://lists.openstack.org/pipermail/openstack-dev/2017-March/113370.html

Change-Id: I91d088d0ac8fb0164cb7616c3fd6022d44f953c8