Updated openstack/openstack Project: openstack/openstack-manuals  38811714e9d6d790c036c3ee182b8278fb10cd69
Use Sphinx 1.5 warning-is-error

With pbr 2.0 and Sphinx 1.5, the setting for treat sphinx warnings as
errors is setting warning-is-error in build_sphinx section. Migrate
the setting from the old warnerrors one.

Remove all-files, we do not have man pages.

Note that the build_sphinx section is not used by our toolchain, still
doesn't harm to use proper setup so that it could be used.

Change-Id: I03403b1ed20c4a4bccb1c03311d3999e12a98416