Updated openstack/openstack Project: openstack/neutron  038597e79439465f557f450edf3d75d15250426e
Use correct retry mechanism in tags

This replaces wrap_db_retry with
neutron.db.api.retry_if_session_inactive() in the tags
plugin.

wrap_db_retry is missing a bunch of exception types we
can encounter and are expected to retry in normal database
operations in Neutron (e.g. staledataerror in the bug report).

This patch doesn't include a unit test because there is already
quite a bit of coverage ensuring the decorator handles these types
of exceptions.

Closes-Bug: #1673243
Change-Id: I65b10ee54301401cd4525152a82a604ac8744582