Updated openstack/openstack Project: openstack/horizon  fc8fae167aeb3d8821142d11023a93c5bd55ae15
The progress bar is wrong when you create a snapshot for a volume.

The progress bar is wrong when you create a snapshot for a volume,
it shows the progress bar for the volumes, not for the snapshots.
Change the template for the snapshot limit in order to show the
right progress bar.

Change-Id: Ibdaff7469c389bd7161ba1466641a623db351ac9
Close-Bug: #1671457