Updated openstack/openstack Project: openstack/tripleo-heat-templates  5b2a1bc3293d02689a70f8f5b35128bf216edd96
Keep existing data for containerized MongoDB

We used named Docker volume for MongoDB storage, which meant that when
moving from bare metal to containerized, we lost data and reinitialized
the storage from scratch.

With this commit we keep the data by mounting the original data into the
container. We also need make sure that file ownership is correct
according to the uid/gid used within MongoDB container image.

Change-Id: I86ef2cb37a068b767462d6d50fe451389b7cbb58