Updated openstack/openstack Project: openstack/horizon  c0cb55046e5612765bc564eb1197cb696e522c16
Make extract_messages quiet in pep8 test

The additional extract_messages call in the pep8 tox environment
generates a lot of meaningless output, that then makes it difficult
to spot actual failures and warnings. This patch makes the Django
command respect the verbosity setting and call the setup.py command
with a quiet flag if low verbosity is requested, considerably decreasing
the amount of useless output. It then sets that verbosity to low in the
corresponding tox environment.

Change-Id: I2976250a23e800fbb582726299670b57199b58a7