Updated openstack/openstack Project: openstack/tripleo-heat-templates  c571ca7b763fee6cb177c107f65bc314daf33ca8
docker/keystone: add metadata_settings to output

This is used for the TLS-everywhere bits. It will be taken into account
by a metadata hook that outputs relevant entries for the nova-metadata
service; and subsequently kerberos principals will be created from
these.

Subsequent patches will add support for TLS in the internal network for
the containerized keystone.

Change-Id: Ic747ad9c8d6e76c8c16e347c1cdcabc899dd9f9a