Updated openstack/openstack Project: openstack/nova-specs  fe6b796e78fba15e472e2b755d77b7c03e8b5a11
Move implemented specs for Ocata

This moves the specs that were completed in Ocata
from the 'approved' directory to the 'implemented'
directory and updates the redirects file.

This also fixes a bug in the move-implemented-specs
script such that we filter out completed blueprints
which are not implemented, which was the case with
blueprint cold-migration-with-target-ocata which was
superseded.

Change-Id: I476e2e2d4bb089d901b887ce7728fc29c35a0957