Updated openstack/openstack Project: openstack-infra/tripleo-ci  bd485839781734fcd543321bd3312834ad6d5e8e
Revert second repo-setup run for overcloud nodes

This commit reverts the workaround that was the second run of repo-setup
on overcloud noeds. This workaround was a short-term fix to allow stable
releases to use CI, however a better fix has been created involving the
refactoring of the vxlan networking code that is dependent on this
patch.

Change-Id: Ibcb7d4e0ef603b94a9b8f59b8f81450919c1a371