Updated openstack/openstack Project: openstack/heat  0afcfaa35c7e5c966ef138471824473f7031bf26
Add a make_url intrinsic function

A large proportion of uses of the str_replace function is to build URLs out
of various components from various sources. This is invariably brittle,
with a failure to escape special characters, deal with IPv6 addresses, and
so on. The make_url function provides a both a tidier interface and correct
handling of these edge cases.

Change-Id: I61b6dff01cd509b3d1c54bca118632c276569f4e