Updated openstack/openstack Project: openstack/cinder  49040c13ab689a40d1cea2f202deab986deb8ed2
Add quota limit check and enhance roll back for cg create

when a consistency group is created from a source consistency group
and the number of volumes associated with the source CG exceeds
the volumes quota, the creation of the CG will fail and will leave
volumes in "creating".

This patch add the quota limit check and also let the new created
volumes can be deleted if error occured.

Change-Id: I4f2ce1999c0eee798e2f4650a940d7335c548d40
Closes-bug: #1609889