Updated openstack/openstack Project: openstack/nova  0b09cee2c34cd197e7909a3251ac697e00d7b50d
compute: Only destroy BDMs after successful detach call

Previously if detach was called with destroy_bdm=True the BDM would
always be destroyed first before any detach call was made to cinder.
This could result in the destruction of a BDM while the volume remained
marked as attached in Cinder.

This has now changed so the BDM is only destroyed after a successful
detach call has been made to cinder. In the event of a failure during
this detach call to Cinder an operator can now make another attempt to
detach the volume via Nova as the BDM is no longer destroyed.

The associated volume.detach notification is also now only sent after a
successful detach call to Cinder and before the BDM is destroyed.

Change-Id: I2581ff9f9c0e7cfc14a25acf45eb1860df69eacf