Updated openstack/openstack Project: openstack/ironic-python-agent  4be702242eb48e09eb77e46d997908b7c519127d
Add missing 'autospec' argument to mock.patch

Add missing 'autospec' keyword argument to mock.patch and
mock.patch.object calls. Use 'autospec=True' except for a few cases
where it fails because the mocked function is a @classmethod and it
doesn't work. In that case explicity set it to 'autospec=False'

Change-Id: I620dce91abaa4440e1803aeefb3e93c0b65d1419