Updated openstack/openstack Project: openstack/heat  7c908e220df4dba9f1845dbc8021e372c6326f73
Remove bandit.yaml in favor of default config

Latest g-r version of bandit does not require a bandit.yaml. This
removes the bandit.yaml and skips the same tests that were not
run as defined by the profile before. This patch also puts running
of bandit within the pep8 testenv, consistent with other projects.

Change-Id: I7db8bc460e2051c635dc7b3fd4b152ece80c7a0c