Updated openstack/openstack Project: openstack-infra/project-config  02d2a2fcffe9baaa2796ae1a2b0dac078d0ce93d
Use StoryBoard for Monasca repositories

The Monasca team has agreed to move their task and defect tracking
into StoryBoard, so all repositories for official Monasca
deliverables get the use-storyboard flag and, if missing, the
monasca project group added. Also do this for the monasca-events-api
repo which seems to have been overlooked in governance (but still
shares the monasca bug project in LP and monasca's Gerrit ACL).

Change-Id: I57434924c7818d859cf5b725094ce397a9c112ca