Updated openstack/openstack Project: openstack-infra/system-config  5599512b27933897eda2d2a60a69fb059eceebd7
Use PYTHONIOENCODING=utf-8 for invites invocation

Since the send.py script for invites uses print to emit progress
including the names of contacts to which messages are sent,
redirecting stdout of the script to a local file can result in
encoding errors if some of the names include non-ASCII codepoints
(but not when stdout is your interactive terminal). Setting
PYTHONIOENCODING=utf-8 in the calling environment works around this
nicely.

Change-Id: I52b70c64f8b3a20da5143b31c5090192b0c75c99