Updated openstack/openstack Project: openstack-infra/project-config  b0f023e3b0d0676a51da63cb2dcd9b22c1a8c17a
Add monitorstack project for OpenStack-Ansible

This patch adds new pre-seeded git repository for a monitoring plugin
framework for OpenStack clouds.

This will be managed and curated by the OpenStack-Ansible core team, but
the end aim will be to move this into it's own group at a point where it
is adopted more widely.

Change-Id: I61c687a1fd89d5ac43b97a56dc1aa33cdd1528ee
Needed-By: Iadbe8ae4e205932f55907dea6931a42b63bf62e7