Updated openstack/openstack Project: openstack/tripleo-heat-templates  e0bd63c826e687d9019b76297e9375f3b0608c2e
Keep existing data for containerized ironic-conductor

Use mounts instead of docker volumes, and preserve existing data when
moving from baremetal to containerized ironic-conductor.

We cannot keep the data in the same directory to avoid hard-linking
errors in ironic, because of this issue:

https://github.com/docker/docker/issues/7457

This means we need to copy the data over to a new location before we
start the containers.

Change-Id: If98460120212f887b06adf117c5d88b97682638e