Updated openstack/openstack Project: openstack/governance  b55e7c82cf8f33af8e62d5b13531986bbf3d6eca
Add sushy under Ironic governance

Sushy is a Python library to communicate with Redfish based systems
(http://www.dmtf.org/standards/redfish) which is currently being used to
create a "redfish" driver/hardware type for Ironic.

In the upstream meeting on 2017-03-13 it was discussed whether we want
to include it under the Ironic's governance or not [0] and the conclusion
is yes, we want it to be part of the governance.

[0] http://eavesdrop.openstack.org/meetings/ironic/2017/ironic.2017-03-13-17.00.log.html#l-362

Change-Id: I8bc6d02653be8c34eea3dbe768ab45a68aed1d11