Updated openstack/openstack Project: openstack/horizon  ccfa52fad10977a54b185936b284b383625291b3
Add charts to show volume quotas on Python launch instance modal

Added two charts for Number of Volumes and Total Volume Storage quotas
on Python launch instance modal when cinder is enabled.

The charts reflect the volume usage of the new instances to be created
as the user changes the configuration on the modal.

Change-Id: I68a88cc867dd36f3d29cd77c1825858f86f3713f
Partially-implements: blueprint launch-instance-volume-quotas