Updated openstack/openstack Project: openstack/requirements  9b03abaf3e3a1fc687187eafc46d9b0c21aaeadf
Add virtualbmc to requirements and constraints

virtualbmc is a library/tool that is used throughout projects under
OpenStack Baremetal program for testing on virtualized hardware.

It depends on pyghmi library (pure Python IPMI protocol implementation)
which is already in g-r and u-c.

In order to avoid incompatible versions of virtualbmc and pyghmi
installed or required, which was already seen (ML thread [0]),
virtualbmc should also be present in g-r and u-c.

Versions specified in this patch are latest currently available and are
known to work on CI already.

[0] http://lists.openstack.org/pipermail/openstack-dev/2017-March/113152.html

Change-Id: I72a170a0ab44673dd3041fe102d3b910e7db203f