Updated openstack/openstack Project: openstack/barbican  e88d79cae0c0b757c25becf8b65c6f078bc176fa
Fix KMIP gate

Switches KMIP server configuration and KMIP plugin to use TLSv1.2
by default. This is the most secure option. In case the system is
older and does not have TLSv1.2, an error message is printed out.

Removes the behavior of switching to TLSv1.2 despite
the user's configuration that was added with
I7018262cb74a95dfa24d6b94d49f1ebd62bdeebd. This behavior was
confusing -- it is much clearer to have the user-configurable options
default to TLSv1.2.

Sets KMIP_PLUGIN_ENABLED before running the tests, so that the
tests that won't work for KMIP will be skipped.
See Id908bf57233af84bff56d90c75d175b04ccd4373 for more details.

Cleans up the quotes around the paths to the log files and server
conf files when they are passed to the pykmip server.

Depends-On: I9fe7b156c4a825c8bfe94a3c48ce686ce0dee01e
Change-Id: I64e27a26dfe02d794b725763c55d0197bc2c46bd