Updated openstack/openstack Project: openstack-infra/shade  ef4279ba0efd54da212b0645b4bb251da7c3f71e
delete_server: make sure we sleep a bit when waiting for server deletion

If we call delete_server(.., wait=True) then self._SERVER_AGE is
the 'sleeping time' between 2 calls to 'get_server'. But by default
self._SERVER_AGE is 0, so we are hitting the Nova server hard.

This patch is to make sure that we wait for some time (2sec) if
self._SERVER_AGE has not been tweaked.

Change-Id: I4aa7b3554348bd0b8fec3069fe8d0e3142e47bb3