Updated openstack/openstack Project: openstack/nova  3ee4b0cf0ebd6db5118d0fcdc6f9f76e2fcd521b
conf: Deprecate 'use_ipv6'

Since Idcfdaf3b removed the use of this flag in the libvirt port filter
and I188fc2c removed the use in injected network templates, the only
remaining use of this option is in nova-network. As a result, we can
deprecate this like nova-network itself.

Change-Id: Ibbcd1fd11c3563b4ddef7c128b714402beac7e3d
Implements: blueprint centralize-config-options-pike