Updated openstack/openstack Project: openstack-infra/project-config  e1ae1ff95375217b86977889264367fc5f1da581
puppet/tripleo: stop testing Puppet 3 in Pike and beyond

Puppet 3 is not supported by Puppet OpenStack community starting from
Pike cycle. So we don't want to run the CI jobs anymore.

Because infra is still running syntax-3 jobs, we create a -legacy job
only used by Puppet OpenStack, and run it only on the branches that we
want.

Change-Id: Ifc742361efce1b7f56a3b23046d72642ba697434